[TS-30] Feat: SelectTimeModal 초기 추가
